Adding the Pointer Icon Sample for Android on CrOS
Adding pointer icons is a quick and easy way to
help your users visually notice when elements
within your Android application are useful. This
can be a way to provide context to the user about
the available actions of items within your
application. The sample aims to provide a quick
way to check which pointer icons look like what
and how to implement a pointer icon on a hover
action within your application.